レスのFormatterクラスを実装
