+++ lib/bbiff.rb
@@ -1,8 +1,10 @@
 require 'shellwords'
 require 'optparse'
-require_relative 'bbiff/version'
-require_relative 'bbiff/bbs_reader'
-require_relative 'bbiff/res_format'
+require 'active_support'
+require 'active_support/core_ext/string'
+
+require 'bbiff/version'
+require 'bbiff/bbs_reader'
 
 def parse_range(str)
   if str == "all"
@@ -18,14 +20,15 @@ def parse_range(str)
   end
 end
 
-def start_polling(thread, start_no, command)
+def start_polling(thread, start_no, command, format_type='default')
   default_notify_command = 'notify-send'
   notify_send = ENV['BBIFF_NOTIFY_SEND'] ||
                 command ||
                 (system("which #{default_notify_command}") ? default_notify_command : 'echo')
+  formatter = eval "Bbiff::Formatter::#{format_type.camelcase}"
   loop do
     thread.posts(parse_range("#{start_no}-")).each do |post|
-      system("#{notify_send} #{Shellwords.escape(render_post(post))}")
+      system("#{notify_send} #{Shellwords.escape(formatter.format(post))}")
       sleep 1
     end
     start_no = thread.last + 1
@@ -49,6 +52,7 @@ Copyright © 2016 Yoteichi
 --
 
     -c, --command COMMAND   Specify notify-send command.
+    -f, --format  FORMAT    Specify res format type.
     -h, --help              Display this help message.
 EOD
 end
@@ -60,10 +64,19 @@ def main
   end
 
   command = nil
+  format_type = 'default'
   opt = OptionParser.new
   opt.on('-h', '--help') { usage; exit }
   opt.on('-c COMMAND', '--command COMMAND') { |v| command = v }
+  opt.on('-f FORMAT', '--format FORMAT') { |v| format_type = v }
   opt.parse!(ARGV)
+
+  begin
+    require "bbiff/formatter/#{format_type}"
+  rescue LoadError => e
+    STDERR.puts "Error: Unknown format [#{format_type}]"
+    exit 1
+  end
 
   url = ARGV[0]
 
@@ -74,5 +87,5 @@ def main
 
   thread = Bbs::C板.new(*ita).thread(sure)
   start_no = ARGV[1] ? ARGV[1].to_i : thread.last + 1
-  start_polling(thread, start_no, command)
+  start_polling(thread, start_no, command, format_type)
 end
